Use mushrrom nikita ppo

+++ atacom/agent_builder/ppo_atacom.py
@@ -1,6 +1,7 @@
 import torch
 
-from atacom.agent_builder.ppo_nikita import NikitaPPO
+
+from mushroom_rl.algorithms.actor_critic.deep_actor_critic.ppo_nikita import NikitaPPO
 from mushroom_rl.rl_utils.value_functions import compute_gae
 
 class AtacomPPO(NikitaPPO):
@@ -29,7 +30,7 @@ class AtacomPPO(NikitaPPO):
 
         old_pol_dist = self.policy.distribution_t(state_old)
 
-        old_log_p = old_pol_dist.log_prob(action.to(dtype=torch.float32))[:, None].detach()
+        old_log_p = old_pol_dist.log_prob(action)[:, None].detach()
 
         self._V.fit(state, v_target, **self._critic_fit_params)
